server/requirements.txt
# MagisAI Training Server dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pyyaml>=6.0
pydantic>=2.0
